Always use babelrc with react rewired commands
For some reason .babelrc isn't recognizing NODE_ENV or BABEL_ENV
so it always defaults to 'development'. I don't know if it's
babel or CRA or react rewired, and I don't feel like digging into
the internals. So, we'll just always use istanbul when running
a script with react rewired, and we'll just build with the
original CRA react-script command.