Update tests to work in TS/Nextjs
Most of the changes were to fix types, but I also had to add a
transformer for svg imports, which I copied from SO of course.
I left ErrorBoundary as a class component, because there isn't a
hook equivalent for componentDidCatch, and try/catch doesn't work,
and this was just easier.