Update GQL types to use tyepscript
This required going back to using yarn, because something about
how NPM resolves subdependencies was creating conflicts inside
Apollo, which was raising errors when I tried to generate the
schema and/or types. I tried the various version-based fixes
suggested in SO & GH issues, but switching back to yarn fixed it
without having to get so picky with nested versions.